First compilable source after refactoring
